docs(api): Move HealthResponse and StatusResponse to api_schemas.py
